Fixlint by moving preview to a new file
