+++ Views/Library/ZimFileDetail.swift
@@ -377,30 +377,3 @@ private struct ServiceWorkerWarning: View {
         }
     }
 }
-
-struct ZimFileDetail_Previews: PreviewProvider {
-    static let context = NSManagedObjectContext(concurrencyType: .mainQueueConcurrencyType)
-    static let zimFile: ZimFile = {
-        let zimFile = ZimFile(context: context)
-        zimFile.articleCount = 1000000
-        zimFile.category = "wikipedia"
-        zimFile.created = Date()
-        zimFile.downloadURL = URL(string: "https://www.example.com")
-        zimFile.fileID = UUID()
-        zimFile.fileDescription = "A very long description"
-        zimFile.flavor = "max"
-        zimFile.hasDetails = true
-        zimFile.hasPictures = false
-        zimFile.hasVideos = true
-        zimFile.languageCode = "en"
-        zimFile.mediaCount = 100
-        zimFile.name = "Wikipedia Zim File Name"
-        zimFile.persistentID = ""
-        zimFile.size = 1000000000
-        return zimFile
-    }()
-
-    static var previews: some View {
-        ZimFileDetail(zimFile: zimFile, dismissParent: nil).frame(width: 300).previewLayout(.sizeThatFits)
-    }
-}
